backend/requirements.txt
flask
flask-cors
pyserial
pandas
scikit-learn
requests
python-dotenv
